refactor: ProgressBoard 컴포넌트 분리 및 최적화
- ProgressBoard를 단일 책임 원칙에 따라 4개 컴포넌트로 분리
  - CollapseButton, ExpandButton, ProgressBar, StageList
